finish correcting mistakes in TournamentTracker

+++ TournamentTracker/ClassLibrary3/Models/ModelMatchupEntry.cs
@@ -1,11 +1,11 @@
 ﻿namespace TrackerLibrary.Models
 {
-    public class ModelMatchupEntry // REFACTORED
+    public class ModelMatchupEntry // x
     {
         public int Id { get; set; }
         public int TeamCompetingId { get; set; }
         public ModelTeam TeamCompeting { get; set; }
-        public double Score { get; set; }
+        public int Score { get; set; }
         public int ParentMatchupId { get; set; }
         public ModelMatchup ParentMatchup { get; set; }
     }
