+++ TournamentTracker/ClassLibrary3/Models/ModelTournament.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 
 namespace TrackerLibrary.Models
 {
-    public class ModelTournament // REFACTORED
+    public class ModelTournament // x
     {
         public event EventHandler<DateTime> OnTournamentComplete;
         public int Id { get; set; }
         public string TournamentName { get; set; }
-        public decimal EntryFee { get; set; }
+        public int EntryFee { get; set; }
         public List<ModelTeam> EnteredTeams { get; set; } = new List<ModelTeam>();
         public List<ModelPrize> Prizes { get; set; } = new List<ModelPrize>();
         public List<List<ModelMatchup>> Rounds { get; set; } = new List<List<ModelMatchup>>();
